% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socratr-package.R
\docType{package}
\name{socratr-package}
\alias{socratr}
\alias{socratr-package}
\title{socratr: Fast R Interface to Socrata Open Data Portals}
\description{
Read, write, and list datasets on Socrata open data portals \url{https://dev.socrata.com}. Supports SODA 3 JSON and SODA 2 CSV downloads with automatic pagination, optional schema-based type coercion, parallel fetching and uploading for large datasets, and Discovery API listing. Inspired by the RSocrata package \url{https://github.com/Chicago/RSocrata} by the City of Chicago.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dschaadt32/socratr}
  \item Report bugs at \url{https://github.com/dschaadt32/socratr/issues}
}

}
\author{
\strong{Maintainer}: Derek Schaadt \email{dschaadt@alum.mit.edu}

}
\keyword{internal}
