% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socratr.R
\name{is_four_by_four}
\alias{is_four_by_four}
\title{Test whether a string is a valid Socrata 4x4 dataset identifier}
\usage{
is_four_by_four(x)
}
\arguments{
\item{x}{Character vector to test.}
}
\value{
Logical vector the same length as `x`.
}
\description{
Test whether a string is a valid Socrata 4x4 dataset identifier
}
\examples{
is_four_by_four("abcd-1234")   # TRUE
is_four_by_four("not-valid!")  # FALSE
}
