% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSocrata.R
\name{posixify}
\alias{posixify}
\title{Convert Socrata calendar_date string to POSIXct}
\usage{
posixify(x)
}
\arguments{
\item{x}{character vector of Socrata dates (CSV or JSON formats)}
}
\value{
a POSIXct vector
}
\description{
Convert Socrata calendar_date string to POSIXct
}
