% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSocrata.R, R/socratr.R
\name{posixify}
\alias{posixify}
\title{Convert Socrata calendar_date string to POSIX}
\usage{
posixify(x)

posixify(x)
}
\arguments{
\item{x}{Character vector of date/time strings.}
}
\value{
a POSIX date

A `POSIXct` vector in the system timezone.
}
\description{
Handles ISO 8601 (SODA 3) and legacy `mm/dd/yyyy` (SODA 2) formats.
Warns if any non-missing value fails to parse.
}
\examples{
posixify("2024-06-15T13:45:00.000")
posixify("06/15/2024 01:45:00 PM")
}
\author{
Hugh J. Devlin, Ph. D. \email{Hugh.Devlin@cityofchicago.org}
}
